Use activities folder entrypoint
Committed-By-Agent: codex
