+++ apps/service/src/api/app.integration.test.ts
@@ -11,7 +11,7 @@ import Stripe from 'stripe'
 import sourceStripe from '@stripe/sync-source-stripe'
 import destinationPostgres from '@stripe/sync-destination-postgres'
 import { createApp as createEngineApp, createConnectorResolver } from '@stripe/sync-engine'
-import { createActivities } from '../temporal/activities.js'
+import { createActivities } from '../temporal/activities/index.js'
 import { createApp } from './app.js'
 import type { paths } from '../__generated__/openapi.js'
