+++ apps/service/src/api/app.test.ts
@@ -5,7 +5,7 @@ import { Worker } from '@temporalio/worker'
 import path from 'node:path'
 import { createConnectorResolver, sourceTest, destinationTest } from '@stripe/sync-engine'
 import destinationGoogleSheets from '@stripe/sync-destination-google-sheets'
-import type { SyncActivities, RunResult } from '../temporal/activities.js'
+import type { SyncActivities, RunResult } from '../temporal/activities/index.js'
 import { createApp } from './app.js'
 
 const resolver = createConnectorResolver({
