+++ apps/service/src/index.ts
@@ -17,8 +17,8 @@ export { createApp } from './api/app.js'
 export type { AppOptions } from './api/app.js'
 
 // Temporal workflow types (for consumers that need to reference them)
-export { createActivities } from './temporal/activities.js'
-export type { SyncActivities, RunResult } from './temporal/activities.js'
+export { createActivities } from './temporal/activities/index.js'
+export type { SyncActivities, RunResult } from './temporal/activities/index.js'
 export type { WorkflowStatus } from './temporal/workflows/_shared.js'
 export { createWorker } from './temporal/worker.js'
 export type { WorkerOptions } from './temporal/worker.js'
